Update ffi interface to match the latest version of the C code

+++ src/channel/builder.rs
@@ -13,7 +13,7 @@
 //  0. You just DO WHAT THE FUCK YOU WANT TO.
 
 use std::{mem, ptr};
-use libc::c_int;
+use libc::{c_int, uint8_t};
 use {ffi, Strip};
 use super::Channel;
 
@@ -41,8 +41,8 @@ impl Builder {
 		self
 	}
 
-	pub fn brightness(&mut self, value: i32) -> &mut Self {
-		self.0.brightness = value as c_int;
+	pub fn brightness(&mut self, value: u8) -> &mut Self {
+		self.0.brightness = value as uint8_t;
 		self
 	}
